Started cleaning up user settings
