+++ src/common/page_components/settings/general.js
@@ -1,5 +1,3 @@
-import { useState, useEffect, useContext } from 'react'
-import Image from 'next/image'
 import { useUser } from '../../lib/hooks'
 
 
@@ -11,43 +9,20 @@ export function General() {
     const [user] = useUser({ userOnly: true })
     const [context, setContext] = useAppContext()
 
-    async function saveUsername(value) {
-        const res = await fetch(window.origin + "/api/user/setUsername", {
-            body: JSON.stringify({
-                username: value
-            }),
-            method: "POST"
-        })
-        const newUser = await res.json()
-        context.data["settings.username"].value = newUser.username
-        context.user = newUser
-        setContext({ ...context })
-    }
-
-    async function saveAccountPublic(value) {
-        const res = await fetch(window.origin + "/api/user/setAccountPublic", {
-            body: JSON.stringify({
-                public: value
-            }),
-            method: "POST"
-        })
-        const newUser = await res.json()
-        context.data["settings.accountPublic"].value = newUser.public == true
-        context.user = newUser
-        setContext({ ...context })
-    }
-
-    async function saveWeeklyHourGoal(value) {
-        const res = await fetch(window.origin + "/api/user/setWeeklyHourGoal", {
-            body: JSON.stringify({
-                goal: value
-            }),
-            method: "POST"
-        })
-        const newUser = await res.json()
-        context.data["settings.weeklyHourGoal"].value = newUser.weeklyHourGoal
-        context.user = newUser
-        setContext({ ...context })
+    function saveSetting(contextKey, key) {
+        return async (data) => {
+            const toSend = {}
+            toSend[key] = data
+            console.log(toSend)
+            const res = await fetch(window.origin + "/api/user/settings", {
+                body: JSON.stringify(toSend),
+                method: "PATCH"
+            })
+            const newUser = await res.json()
+            context.data[contextKey].value = newUser[key]
+            context.user = newUser
+            setContext({ ...context })
+        }
     }
 
     async function saveProfilePicture(value) {
@@ -68,17 +43,17 @@ export function General() {
 
             <div className={styles.entry}>
                 <div className={styles.key}>username</div>
-                <FormInput width={"40%"} type="text" defaultValue={user.getUsername()} onSave={saveUsername} contextKey="settings.username" />
+                <FormInput width={"40%"} type="text" defaultValue={user.getUsername()} onSave={saveSetting("settings.username", "username")} contextKey="settings.username" />
             </div>
 
             <div className={styles.entry}>
                 <div className={styles.key}>weekly hour goal</div>
-                <FormInput width={"40%"} type="number" defaultValue={user.getWeeklyHourGoal()} onSave={saveWeeklyHourGoal} contextKey="settings.weeklyHourGoal" />
+                <FormInput width={"40%"} type="number" defaultValue={user.getWeeklyHourGoal()} onSave={saveSetting("settings.weeklyHourGoal", "weeklyHourGoal")} contextKey="settings.weeklyHourGoal" />
             </div>
 
             <div className={styles.entry}>
                 <div className={styles.key}>account public</div>
-                <FormInput type="checkbox" defaultValue={user.getAccountPublic()} onSave={saveAccountPublic} contextKey="settings.accountPublic" />
+                <FormInput type="checkbox" defaultValue={user.getAccountPublic()} onSave={saveSetting("settings.accountPublic", "public")} contextKey="settings.accountPublic" />
             </div>
 
             <div className={styles.entry}>
