--- src/pages/api/user/setAccountPublic.js
@@ -1,21 +1,0 @@
-import { getUser } from './index'
-
-async function Route(req, res) {
-    await getUser(req, {
-        onFound: async user => {
-            req.body = JSON.parse(req.body)
-
-            const accountPublic = req.body.public == true
-
-            user.public = accountPublic
-
-            await user.save()
-            res.status(200).json(user)
-        },
-        onNotFound: () => {
-            res.status(500).end('Authentication token is invalid, please log in')
-        }
-    })
-}
-
-export default Route
