--- src/pages/api/user/setWeeklyHourGoal.js
@@ -1,22 +1,0 @@
-import { getUser } from './index'
-
-async function Route(req, res) {
-    await getUser(req, {
-        onFound: async user => {
-            req.body = JSON.parse(req.body)
-
-            const newGoal = req.body.goal
-
-            if (!isNaN(newGoal))
-                user.weeklyHourGoal = newGoal
-
-            await user.save()
-            res.status(200).json(user)
-        },
-        onNotFound: () => {
-            res.status(500).end('Authentication token is invalid, please log in')
-        }
-    })
-}
-
-export default Route
